refactor(faq): extract FAQ_ITEMS into lib/faq.ts
Decouples the root layout's JSON-LD generation from the FAQ UI module
so app/layout.tsx no longer transitively imports Frame/SectionHeader/FaqItem
just to read the data.
